+++ test/beast/websocket/write.cpp
@@ -287,7 +287,7 @@ public:
                         BOOST_THROW_EXCEPTION(
                             system_error{ec});
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             BEAST_EXPECT(count == 0);
             ws.async_write(sbuf("*"),
                 [&](error_code ec, std::size_t n)
@@ -320,7 +320,7 @@ public:
                         BOOST_THROW_EXCEPTION(
                             system_error{ec});
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             BEAST_EXPECT(count == 0);
             ws.async_write(sbuf("*"),
                 [&](error_code ec, std::size_t)
@@ -356,7 +356,7 @@ public:
                         BOOST_THROW_EXCEPTION(
                             system_error{ec});
                 });
-            while(! ws.wr_block_)
+            while(! ws.wr_block_.is_locked())
             {
                 ioc.run_one();
                 if(! BEAST_EXPECT(! ioc.stopped()))
@@ -398,7 +398,7 @@ public:
                             system_error{ec});
                     BEAST_EXPECT(n == 16384);
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             ws.async_ping("",
                 [&](error_code ec)
                 {
@@ -432,7 +432,7 @@ public:
                             system_error{ec});
                     BEAST_EXPECT(n == 16384);
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             ws.async_ping("",
                 [&](error_code ec)
                 {
@@ -466,7 +466,7 @@ public:
                             system_error{ec});
                     BEAST_EXPECT(n == 16384);
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             ws.async_ping("",
                 [&](error_code ec)
                 {
@@ -499,7 +499,7 @@ public:
                             system_error{ec});
                     BEAST_EXPECT(n == 16384);
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             ws.async_ping("",
                 [&](error_code ec)
                 {
@@ -537,7 +537,7 @@ public:
                             system_error{ec});
                     BEAST_EXPECT(n == s.size());
                 });
-            BEAST_EXPECT(ws.wr_block_);
+            BEAST_EXPECT(ws.wr_block_.is_locked());
             ws.async_ping("",
                 [&](error_code ec)
                 {