+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 155:
 
 * Fix memory leak in advanced server examples
+* Fix soft-mutex assert in websocket stream
 
 --------------------------------------------------------------------------------